backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
SQLAlchemy
psycopg2-binary
pytest 
httpx